ReadMe
this project is repesenting  a game  .(Packmen) using Gui and using DGraph project. (https://github.com/YairAn/Ex2.git) 
in our project we are taking all the informtuion from Jsons file ,  ech repesent a level in the game.
for each level the jason file give the informtion for the fruits and the robots in the level. and graph (Nodes+Edges).

The robots travel on the graph in main to catch the red the fruits as fast as they can to gain points, The more you catch the speed of the robot is increase.

There are two modes for the game:
Automatic game - our algorythem for getting the best score.locating the robot in the best loction in the graph and finding the closest and the high score fruits.
Manual game you choose the robots loction ,and move them with the mouse.
In the end of any game .all the informion about the robot ans fruits will save and there to export the game in to an KML file.